+++ docs/whatsnew.md
@@ -7,6 +7,7 @@
 - fix : correct repr of Variable (VT was incorrect)
 - fix : bug reading SNMP variables
 - new property 'isStructure' for Variable object
+- new members 'toIEC' and 'fromIEC' for Variable object
 - new property 'cpuInfoExtended' for Cpu object
 - new member function 'readRequestStatus' for all objects